[FIX] account: Error when duplicating a purchase invoice
In multi company, when a user 1 creates a purchase invoice I in company A.
The responsible of the invpoice I is user 1. But if the user 1 changes his company from
company A to company B. Then it raised an access error when a user 2 in company A
tries to duplicate the invoice I because the responsible is in an other company.

Now the responsible is not duplicated to avoid access error.

opw:805573