[FIX] web: traceback when exporting from list view
Former code of list controller used the searchView of its parent,
which was the ViewManager. However, this has been changed by rev.
c1ab59237, where the ViewManager was removed, and the list
controller now has the reference of the searchView, so this code
had to be adapted. Tests have been adapted accordingly as well.