[FIX] google_drive: tests: hasSidebar param
Rev. d3eddf8b fixes the google_drive override of the Sidebar in
saas-16. The test specifies param 'sidebar: true', so that a
sidebar is instantiated by the FormView. This param has been
renamed to 'hasSidebar' by rev. c1ab592 in master, so the test
broke when rev. d3eddf8b has been forwardported to master, and it
thus needed to be adapted.