[FIX] hr_timesheet: allow duplication of project
When duplicating a project, a command [(6, _, ids)] may be generated.

This fail with the commit 094f7e8fe9 because all comodels records with
id not in `ids` will have their inverse field unset.

But eg. in the case of copy, this would trigger an error because this
would be done to an empty recordset and unexpectedly trigger an error.

opw-807655
closes #22525