[FIX] mail: livechat: send message with chat_window
Revision on 9dd017236756da84862c0794392b66e78f43b57a
The changes on chat_window are wrong, because the
composer uses an 'input', instead of a 'div'.

As a result, it was not possible to send messages
over im_livechat from the website.

This was not the case on the webclient: chat_window
is enhanced with extended_chat_window, which replaces
the composer with 'div', hence the changes are correct.